Feat: 이벤트 등록 페이지 url 변경
